Remove deprecated visualization functions (#130)
Remove old visualization functions deprecated since #58.

* Remove old tests - fix paths - support pathlib.Path

* fix doc

* move tests/test_visualization_new_structure.py

* cleanup

* Fix broken links
